Style: prettie e lint executado
