+++ .prettierrc
@@ -1,4 +1,8 @@
 {
+  "trailingComma": "all",
   "singleQuote": true,
-  "trailingComma": "all"
+  "endOfLine": "lf",
+  "tabWidth": 4,
+  "arrowParens": "avoid",
+  "printWidth": 80
 }
