+++ src/app.module.ts
@@ -8,29 +8,29 @@ import { ShoppingListModule } from './module/shopping-list/shopping-list.module'
 import { UsersModule } from './module/users/users.module';
 
 const modules = [
-  PersonModule,
-  UsersModule,
-  ProductModule,
-  ShoppingListModule,
-  AuthModule,
+    PersonModule,
+    UsersModule,
+    ProductModule,
+    ShoppingListModule,
+    AuthModule,
 ];
 
 @Module({
-  imports: [
-    ConfigModule.forRoot(),
-    TypeOrmModule.forRoot({
-      type: process.env.DATABASE_TYPE,
-      host: process.env.LOCALHOST_DATABASE,
-      port: process.env.PORT_DATABASE,
-      username: process.env.USERNAME_DATABASE,
-      password: process.env.PASSWORD_DATABASE,
-      database: process.env.DATABASE,
-      entities: [__dirname + '/**/*.entity{.js,.ts}'],
-      synchronize: Boolean(process.env.SYNCHRONIZE_DATABASE),
-    } as TypeOrmModuleOptions),
-    ...modules,
-  ],
-  controllers: [],
-  providers: [],
+    imports: [
+        ConfigModule.forRoot(),
+        TypeOrmModule.forRoot({
+            type: process.env.DATABASE_TYPE,
+            host: process.env.LOCALHOST_DATABASE,
+            port: process.env.PORT_DATABASE,
+            username: process.env.USERNAME_DATABASE,
+            password: process.env.PASSWORD_DATABASE,
+            database: process.env.DATABASE,
+            entities: [__dirname + '/**/*.entity{.js,.ts}'],
+            synchronize: Boolean(process.env.SYNCHRONIZE_DATABASE),
+        } as TypeOrmModuleOptions),
+        ...modules,
+    ],
+    controllers: [],
+    providers: [],
 })
 export class AppModule {}
