+++ src/controllers/auth/auth.controller.ts
@@ -7,13 +7,13 @@ import { AuthService } from '../../services/auth/auth.service';
 @Controller('auth')
 @ApiTags('Auth')
 export class AuthController {
-  constructor(private readonly authService: AuthService) {}
-  
-  @UseGuards(AuthGuard('local'))
-  @ApiBody({ type: LoginDto })
-  @Post('login')
-  @ApiOperation({ summary: 'Login.' })
-  async login(@Req() req) {
-    return await this.authService.login(req.user);
-  }
+    constructor(private readonly authService: AuthService) {}
+
+    @UseGuards(AuthGuard('local'))
+    @ApiBody({ type: LoginDto })
+    @Post('login')
+    @ApiOperation({ summary: 'Login.' })
+    async login(@Req() req) {
+        return await this.authService.login(req.user);
+    }
 }
