+++ src/controllers/auth/strategies/jwt.strategy.ts
@@ -4,15 +4,15 @@ import { ExtractJwt, Strategy } from 'passport-jwt';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor() {
-    super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      ignoreExpiration: false,
-      secretOrKey: process.env.JWT_SECRET_KEY,
-    });
-  }
+    constructor() {
+        super({
+            jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+            ignoreExpiration: false,
+            secretOrKey: process.env.JWT_SECRET_KEY,
+        });
+    }
 
-  async validate(payload) {
-    return { id: payload.sub, email: payload.email };
-  }
+    async validate(payload) {
+        return { id: payload.sub, email: payload.email };
+    }
 }
