+++ src/dto/login/login.dto.ts
@@ -2,13 +2,13 @@ import { ApiProperty } from '@nestjs/swagger';
 import { IsEmail, IsNotEmpty, IsString } from 'class-validator';
 
 export class LoginDto {
-  @ApiProperty()
-  @IsEmail()
-  @IsNotEmpty()
-  email: string;
+    @ApiProperty()
+    @IsEmail()
+    @IsNotEmpty()
+    email: string;
 
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  password: string;
+    @ApiProperty()
+    @IsString()
+    @IsNotEmpty()
+    password: string;
 }
