+++ src/dto/person/person.dto.ts
@@ -2,23 +2,23 @@ import { ApiProperty } from '@nestjs/swagger';
 import { IsNotEmpty, IsString } from 'class-validator';
 
 export default class PersonDto {
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  name: string;
+    @ApiProperty()
+    @IsString()
+    @IsNotEmpty()
+    name: string;
 
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  city: string;
+    @ApiProperty()
+    @IsString()
+    @IsNotEmpty()
+    city: string;
 
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  uf: string;
+    @ApiProperty()
+    @IsString()
+    @IsNotEmpty()
+    uf: string;
 
-  @ApiProperty()
-  @IsString()
-  @IsNotEmpty()
-  zipcode: string;
+    @ApiProperty()
+    @IsString()
+    @IsNotEmpty()
+    zipcode: string;
 }
