+++ src/dto/shoppingList/shoppingListUpdate.dto.ts
@@ -2,8 +2,10 @@ import { ApiProperty, OmitType } from '@nestjs/swagger';
 import { IsNumber } from 'class-validator';
 import { ShoppingListDto } from './shoppingList.dto';
 
-export class ShoppingListUpdateDto extends OmitType(ShoppingListDto,['total']) {
-  @ApiProperty()
-  @IsNumber()
-  id: number;
+export class ShoppingListUpdateDto extends OmitType(ShoppingListDto, [
+    'total',
+]) {
+    @ApiProperty()
+    @IsNumber()
+    id: number;
 }
