+++ src/dto/user/userUpdate.dto.ts
@@ -3,12 +3,12 @@ import { IsNotEmpty, IsString, Matches } from 'class-validator';
 import { RegexHelper } from '../../util/regex';
 
 export class UserUpdateDto {
-  @ApiProperty()
-  @IsString()
-  @Matches(RegexHelper.password, {
-    message:
-      'A senha deve conter letras maiúsculas minúsculas, números e caracteres especiais',
-  })
-  @IsNotEmpty()
-  password: string;
+    @ApiProperty()
+    @IsString()
+    @Matches(RegexHelper.password, {
+        message:
+            'A senha deve conter letras maiúsculas minúsculas, números e caracteres especiais',
+    })
+    @IsNotEmpty()
+    password: string;
 }
