+++ src/entities/product.entity.ts
@@ -1,45 +1,45 @@
 import {
-  Column,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  PrimaryGeneratedColumn,
+    Column,
+    Entity,
+    JoinColumn,
+    ManyToOne,
+    PrimaryGeneratedColumn,
 } from 'typeorm';
 import { ShoppingListE } from './shoppingList.entity';
 
 @Entity({ name: 'product' })
 export class ProductE {
-  @PrimaryGeneratedColumn()
-  id?: number;
+    @PrimaryGeneratedColumn()
+    id?: number;
 
-  @Column()
-  store: string;
+    @Column()
+    store: string;
 
-  @Column()
-  category: string
+    @Column()
+    category: string;
 
-  @Column()
-  productName: string;
+    @Column()
+    productName: string;
 
-  @Column()
-  brand: string;
+    @Column()
+    brand: string;
 
-  @Column({ type: 'float' })
-  price: number;
+    @Column({ type: 'float' })
+    price: number;
 
-  @Column()
-  buyDate: string;
+    @Column()
+    buyDate: string;
 
-  @Column({ type: 'float' })
-  unity: number;
+    @Column({ type: 'float' })
+    unity: number;
 
-  @Column()
-  commercialUnit: string;
+    @Column()
+    commercialUnit: string;
 
-  @Column({ type: 'boolean' })
-  wasAcquired: boolean;
+    @Column({ type: 'boolean' })
+    wasAcquired: boolean;
 
-  @ManyToOne(() => ShoppingListE, (list) => list)
-  @JoinColumn({ name: 'list' })
-  list: ShoppingListE;
+    @ManyToOne(() => ShoppingListE, list => list)
+    @JoinColumn({ name: 'list' })
+    list: ShoppingListE;
 }
