+++ src/module/users/users.module.ts
@@ -5,9 +5,9 @@ import { UserController } from '../../controllers/user/user.controller';
 import { UserService } from '../../services/user/user.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([UserE])],
-  controllers: [UserController],
-  providers: [UserService],
-  exports: [UserService],
+    imports: [TypeOrmModule.forFeature([UserE])],
+    controllers: [UserController],
+    providers: [UserService],
+    exports: [UserService],
 })
 export class UsersModule {}
