+++ src/services/auth/auth.service.ts
@@ -6,30 +6,30 @@ import { UserService } from '../user/user.service';
 
 @Injectable()
 export class AuthService {
-  constructor(
-    private readonly userService: UserService,
-    private readonly jwtService: JwtService,
-  ) {}
+    constructor(
+        private readonly userService: UserService,
+        private readonly jwtService: JwtService,
+    ) {}
 
-  async login(user): Promise<object> {
-    const payload = { sub: user.id, email: user.email };
+    async login(user): Promise<object> {
+        const payload = { sub: user.id, email: user.email };
 
-    return {
-      token: this.jwtService.sign(payload),
-    };
-  }
-
-  async validateUser(email: string, password: string): Promise<UserE> {
-    let user: UserE;
-    try {
-      user = await this.userService.findOneOrFail({ where: { email } });
-    } catch (error) {
-      return null;
+        return {
+            token: this.jwtService.sign(payload),
+        };
     }
 
-    const isPasswordValid = compareSync(password, user.password);
-    if (!isPasswordValid) return null;
+    async validateUser(email: string, password: string): Promise<UserE> {
+        let user: UserE;
+        try {
+            user = await this.userService.findOneOrFail({ where: { email } });
+        } catch (error) {
+            return null;
+        }
 
-    return user;
-  }
+        const isPasswordValid = compareSync(password, user.password);
+        if (!isPasswordValid) return null;
+
+        return user;
+    }
 }
